Move printing stuff into its own lib folder
With an eye towards generating a gulp task using a shared lib.

OH and add "singleVerse" option
